requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
blinker=1.8.2=pypi_0
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.7.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
distro=1.9.0=pypi_0
exceptiongroup=1.2.2=pypi_0
faiss-cpu=1.8.0.post1=pypi_0
flask=3.0.3=pypi_0
flask-cors=4.0.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
idna=3.7=pypi_0
importlib-metadata=8.2.0=pypi_0
iniconfig=2.0.0=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markupsafe=2.1.5=pypi_0
mysql-connector-python=9.0.0=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.26.4=pypi_0
openai=1.37.1=pypi_0
openai-async=0.0.3=pypi_0
openssl=3.0.14=h5eee18b_0
packaging=24.1=pypi_0
pip=24.0=py39h06a4308_0
pluggy=1.5.0=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pytest=8.3.2=pypi_0
python=3.9.18=h955ad1f_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
setuptools=69.5.1=py39h06a4308_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
tk=8.6.14=h39e8969_0
tomli=2.0.1=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.4=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=pypi_0
websockets=12.0=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py39h06a4308_0
xz=5.4.6=h5eee18b_1
zipp=3.19.2=pypi_0
zlib=1.2.13=h5eee18b_1
